기능: 한번의 네트워크 통신마다 redis에 기록
