+++ src/main/java/com/bookie/scrap/common/redis/RedisValueNamespace.java
@@ -1,0 +1,17 @@
+package com.bookie.scrap.common.redis;
+
+public enum RedisValueNamespace {
+    Connection("connect:stat"),
+    ;
+
+    private final String prefix;
+
+    RedisValueNamespace(String prefix) {
+        this.prefix = prefix;
+    }
+
+    public String getPrefix() {
+        return prefix;
+    }
+}
+
