+++ src/test/java/com/bookie/scrap/common/redis/EmbeddedRedisTestConfig.java
@@ -1,0 +1,21 @@
+package com.bookie.scrap.common.redis;
+
+import org.springframework.boot.test.context.TestConfiguration;
+import org.springframework.context.annotation.Bean;
+import org.springframework.data.redis.connection.RedisConnectionFactory;
+import org.springframework.data.redis.connection.RedisStandaloneConfiguration;
+import org.springframework.data.redis.connection.lettuce.LettuceConnectionFactory;
+
+@TestConfiguration
+public class EmbeddedRedisTestConfig {
+    @Bean
+    public RedisConnectionFactory redisConnectionFactory() {
+        RedisStandaloneConfiguration config = new RedisStandaloneConfiguration();
+        config.setHostName("fog.torpedo.co.kr");
+        config.setPort(6379);
+        config.setPassword("qwerty");
+
+        return new LettuceConnectionFactory(config);
+    }
+
+}
